server/requirements.txt
python-socketio==5.11.3
fastapi==0.112.1